moved AI chat button to make it appear on labs only. Sulution provided by Eamonn de Leastar
